refactor : 멘토 페이지 클라이언트로 수정했습니다
